auth-service/requirements.txt
Django>=5.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
PyJWT>=2.8.0
requests>=2.31.0